requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
colorama==0.4.6
Django==4.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
loguru==0.7.2
mysqlclient==2.2.4
PyJWT==2.9.0
redis==5.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
win32-setctime==1.1.0
